MUD source files from tape
These are different to the ones that Richard Bartle provided to Stamford. I think these ones are older and dated 1-Mar-84, but reflect the actual MUD as compiled on the tape. These will compile using the enclosed MUD.MIC do file, but you will need to tweak it for your system configuration, particularly the location of BCPL and that Monitor 7.04 no longer supports the NSAVE and NSSAVE commands (as this is the only way to save in Mon 7.04).