Corrected .HIGH. and minor errors
Corrected the .HIGH. values in MUD.MIC for the linker, so that the application puts the right data in high memory and DBASE can produce a working MUD. Fixed text errors in several game .GET files so that DBASE can parse them correctly. Added .DBA files which are required for certain commands in game.