Mud86 case sensitivity
This is the renaming and general tidy up of filenames for the case insensitivity of TOPS-10 vs. the case sensitivity of modern oses. This caused issues in that the repo contained both mud.mic and MUD.MIC, but that MUD.MIC was from the 84 tape and did not work with the rest of this 86 source. All filenames have now been changed to upper case and the Mud84 sources have been removed through the rollback of the earlier commit of POWER.BCL (which doesnt work with Mud86), MUD.MIC (which duplicated mud.mic but also would not compile Mud86), and MBOOTS.MIC which overwrote the Mud86 vesrion.